app/requirements.txt
Flask
requests
opentelemetry-api
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-exporter-otlp
pyroscope-io
pyroscope-otel
